Add files via upload
sorry it's not pretty updated... This add-on should work with the inventory aspects of the array. However the performance metrics of the array is still a work in progress. If you'd like to collaborate, please message me here. I'm more than happy to share int order to find a solution for a Splunk Add-on!